Apply critical fixes
