+++ requirements.txt
@@ -11,3 +11,4 @@ aiofiles>=23.0.0
 pytest>=7.0.0
 pytest-asyncio>=0.21.0
 numpy>=1.24.0
+nest-asyncio>=1.5
